OpenMetadata/requirements.txt
openmetadata-ingestion==1.6.1

jsonschema==4.21.1
requests==2.31.0
google-api-core==2.17.1
